Fix the bug and simplify the MCP example
